- minor fix for eventlog under NIX*

+++ src/zabbix_agent/eventlog.c
@@ -211,7 +211,7 @@ static long    zbx_get_eventlog_message(
 
 int process_eventlog(
 	char *source,
-	int *lastlogsize, 
+	long *lastlogsize, 
 	char *timestamp, 
 	char *src, 
 	char *severity,
